Update the newlinify macro to not externalize its environment
The newlinify macro was externalizing its environment unnecessarily.
Though not necessarily a bug, it made the code both less efficient
and more fragile to the Lisp implementation (there is no guarantee
that environments _be_ externalizable!). Dave Cooper noticed the problem
in ACL 9.0 (where there had been changes in the environment
implementation). Now it's been improved and so no longer a problem.